refactor: remove customelement defined detecting logic (#247)
Before this commit, for those element with tag without `-`, we always
try to detect if the `x-${tagName}` is defined.

After this commit, we pre-define a map(could be override by the
`overrideLynxTagToHTMLTagMap`) to make that transformation for tag name.

This change is a path to SSR and the MTS support.

#52 #54 #51