ci: cleanup CI workflow (#262)
<!--
  Thank you for submitting a pull request!

We appreciate the time and effort you have invested in making these
changes. Please ensure that you provide enough information to allow
others to review your pull request.

Upon submission, your pull request will be automatically assigned with
reviewers.

If you want to learn more about contributing to this project, please
visit:
https://github.com/lynx-family/lynx-stack/blob/main/CONTRIBUTING.md.
-->

## Summary

<!-- Can you explain the reasoning behind implementing this change? What
problem or issue does this pull request resolve? -->

The `.github/workflows/pre-code-review-check.yml` is deleted. And the
`code-style-check` is moved to `test.yml` for better navigation
experience.

Also, a `done` job is added to `test.yml` so that the newly added jobs
do not need to be marked as **REQUIRED** on GitHub.

<!-- It would be helpful if you could provide any relevant context, such
as GitHub issues or related discussions. -->

## Checklist

<!--- Check and mark with an "x" -->

- [x] Tests updated (or **not required**).
- [x] Documentation updated (or **not required**).