feat(source-maps): Add `debugIds` option (#342)
<!--
  Thank you for submitting a pull request!

We appreciate the time and effort you have invested in making these
changes. Please ensure that you provide enough information to allow
others to review your pull request.

Upon submission, your pull request will be automatically assigned with
reviewers.

If you want to learn more about contributing to this project, please
visit:
https://github.com/lynx-family/lynx-stack/blob/main/CONTRIBUTING.md.
-->

## Summary

<!-- Can you explain the reasoning behind implementing this change? What
problem or issue does this pull request resolve? -->

<!-- It would be helpful if you could provide any relevant context, such
as GitHub issues or related discussions. -->

I've noticed that `rspeedy` is using SourceMapDevToolPlugin to generate
source maps, but it doesn't expose the `debugIds` option. This option is
helpful when debugging production errors as it links correct source map
to the correct bundle. More at
https://github.com/tc39/ecma426/blob/main/proposals/debug-id.md

I've tried the change locally with a new Lynx project and the debugIds
comments are generated as expected above existing `sourcemappingurl`.

I'm curious if you would be open to add this option and what needs to be
done to do so.

## Checklist

<!--- Check and mark with an "x" -->

- [x] Tests updated (or not required).
- [x] Documentation updated (or not required).

---------

Signed-off-by: Qingyu Wang <40660121+colinaaa@users.noreply.github.com>
Co-authored-by: Qingyu Wang <40660121+colinaaa@users.noreply.github.com>