ci: remove duplicated playwright test on deploy (#646)
<!--
  Thank you for submitting a pull request!

We appreciate the time and effort you have invested in making these
changes. Please ensure that you provide enough information to allow
others to review your pull request.

Upon submission, your pull request will be automatically assigned with
reviewers.

If you want to learn more about contributing to this project, please
visit:
https://github.com/lynx-family/lynx-stack/blob/main/CONTRIBUTING.md.
-->

## Summary

<!-- Can you explain the reasoning behind implementing this change? What
problem or issue does this pull request resolve? -->

We are removing the `playwright-linux` test from the deployment process,
as it has already been executed in the merge queue.

Note that we are using the `override_branch` option of the
`codecov/test-results-action` to upload the test results. So that we
could still know which tests are flaky.

<!-- It would be helpful if you could provide any relevant context, such
as GitHub issues or related discussions. -->

## Checklist

<!--- Check and mark with an "x" -->

- [ ] Tests updated (or not required).
- [ ] Documentation updated (or not required).

Signed-off-by: Qingyu Wang <40660121+colinaaa@users.noreply.github.com>